pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semquant"
version = "0.1.0"
description = "Automated SEM image quantification (porosity, particle & fiber diameters) with ImageJ-like outputs."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "SEMQuant contributors"}
]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "scikit-image>=0.21",
  "matplotlib>=3.7",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "reportlab>=4.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.28"]
dev = ["pytest>=7.4", "ruff>=0.4", "black>=24.0"]

[project.scripts]
semquant = "semquant.cli:main"

[tool.setuptools]
packages = ["semquant", "semquant.configs"]
include-package-data = true

[tool.setuptools.package-data]
"semquant.configs" = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
